refactor(pr06): rename PhysicalItemDnd35e.mts -> PhysicalItem.mts (G5a)
File-only rename; class is already named PhysicalItem. Updates sibling index.mts.
