+++ src/entities/items/components/Physical/PhysicalItem.mts
